[refactor] sensor twin can be configured to synchronize data with backend
